refactor: split utils + extract TechTree forward-groups (C8, C9)
- **C8**: Extracted `mergeRouteMeta` + `RouteMetaOverrides` + the
  `SITE_URL` constant from `app/utils/utils.tsx` into a focused
  `app/utils/meta.ts`. All 6 route consumers now import the helper
  from `~/utils/meta`. `utils.tsx` drops from ~430 → ~340 lines and
  keeps a clearer charter (layout, dates, skills derivations) — meta
  tags get their own module.
- **C9**: Moved the inline `FORWARD_GROUPS` array + its `ForwardGroup`
  type out of `app/components/TechTree/index.tsx` into a sibling
  `forward-groups.ts`. Component is now purely render logic. Kept the
  data co-located with the component rather than promoting it to
  `skills.json` — these are aspirations (no SKILLS entry, no range),
  not job experiences, so the schema shape wouldn't fit.

Pure refactors — no behaviour change.
