+++ app/routes/education.$slug/index.tsx
@@ -5,7 +5,8 @@ import { FormattedMessage, useIntl } from 'react-intl';
 import Card from '~/components/Card';
 import { loadEducation } from '~/data/education-schema';
 import { type Locale, pickLocale } from '~/intl';
-import { formatDate, getClassMaker, localized, mergeRouteMeta } from '~/utils/utils';
+import { mergeRouteMeta } from '~/utils/meta';
+import { formatDate, getClassMaker, localized } from '~/utils/utils';
 
 import educationJson from '../../../public/data/education.json';
 import styles from './style.css?url';
